requirements.txt
absl-py==0.6.1
astor==0.7.1
boto3==1.9.62
botocore==1.12.62
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
docutils==0.14
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
idna==2.7
imageai==2.0.2
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
protobuf==3.6.1
ptvsd==4.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
scipy==1.1.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
